libopkg: Split out config finalization from config loading
opkg_conf_load() currently loads the configuration from a file and then
proceeds to finalize the config structure by filling in default values
or building the final config values based on other config parameters
(e.g. the offline root).

To enable overwriting of config options from the command line the
finalization phase is split out so it can be called only after the
command line arguments have been processed for the second time.

To retain API compatibility, opkg_conf_load() continues to read and
finalize the config. If the staged approach is desired opkg_conf_read()
followed by a opkg_conf_finalize() shall be used.

+++ libopkg/opkg_conf.c
@@ -707,10 +707,20 @@ int opkg_unlock()
 
 int opkg_conf_load(void)
 {
+    int r = 0;
+
+    r = opkg_conf_read();
+    if (r)
+        return r;
+
+    return opkg_conf_finalize();
+}
+
+int opkg_conf_read(void)
+{
     unsigned int i;
     int r;
     int glob_ret;
-    char *tmp, *tmp_dir_base;
     glob_t globbuf;
     char *etc_opkg_conf_pattern;
 
@@ -768,6 +778,17 @@ int opkg_conf_load(void)
 
         globfree(&globbuf);
     }
+
+    return 0;
+err:
+    opkg_conf_free();
+    return -1;
+}
+
+int opkg_conf_finalize(void)
+{
+    int r;
+    char *tmp, *tmp_dir_base;
 
     /* Option not available on the internal solver since it currently
      * can't merge transactions, resulting in one solve per operation */
@@ -899,17 +920,16 @@ int opkg_conf_load(void)
 
     r = resolve_pkg_dest_list();
     if (r != 0)
-        goto err1;
+        goto err;
 
     nv_pair_list_deinit(&opkg_config->tmp_dest_list);
-
     return 0;
-
- err1:
-    r = rmdir(opkg_config->tmp_dir);
-    if (r == -1)
-        opkg_perror(ERROR, "Couldn't remove dir %s", opkg_config->tmp_dir);
- err:
+err:
+    if (opkg_config->tmp_dir) {
+        r = rmdir(opkg_config->tmp_dir);
+        if (r == -1)
+            opkg_perror(ERROR, "Couldn't remove dir %s", opkg_config->tmp_dir);
+    }
     opkg_conf_free();
     return -1;
 }
